modified:   ebiose_clgp/model/model_config.yaml 	modified:   ebiose_clgp/model/text_encoders/transformer.py

+++ ebiose_clgp/model/text_encoders/transformer.py
@@ -96,13 +96,13 @@ class Transformer(nn.Module):
         nn.init.normal_(self.positional_embedding, std=0.01)
 
     def forward(self, text_data):
-        x = self.token_embedding(text_data).type(self.dtype)  # [batch_size, n_ctx, d_model]
-        x = x + self.positional_embedding.type(self.dtype)
+        x = self.token_embedding(text_data) # [batch_size, n_ctx, d_model]
+        x = x + self.positional_embedding
         x = x.permute(1, 0, 2)  # NLD -> LND
         x = self.resblocks(x)
         x = x.permute(1, 0, 2)  # LND -> NLD
-        x = self.ln_final(x).type(self.dtype)
-
+        x = self.ln_final(x)
+        
         # x.shape = [batch_size, n_ctx, transformer.width]
         # Take features from the eot embedding (eot_token is the highest number in each sequence)
         x = x[torch.arange(x.shape[0]), text_data.argmax(dim=-1)] @ self.text_projection
